refactor: feature-based folder structure
