+++ client/src/components/Users/Auth/SignUpCitizen/StepFour.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { Avatar, Box, Button, Grid, TextField, Typography } from '@mui/material';
 
-import { useStyles } from './styles';
+import { useStyles } from './styles/styles';
 
 interface StepFourType {
   initData: { bio: string };
