+++ client/src/components/Users/Auth/SignUpCitizen/assets/SvgSignUpContactInfoStep.js
