+++ client/src/components/Users/Auth/SignUpCitizen/assets/SvgSignUpFinishedStep.js
