+++ client/src/components/Users/Auth/SignUpCitizen/assets/SvgSignUpInterestsStep.js
