+++ client/src/components/Users/Auth/SignUpCitizen/assets/SvgSignUpLocationStep.js
