+++ client/src/components/Users/Auth/SignUpCitizen/assets/SvgSignUpProfileStep.js
