+++ client/src/components/Users/Auth/SignUpCitizen/constants/interests.ts
